reworked weapon script and created new projectile weapon sub type
