+++ Assets/Prefabs/Projectiles/ClusterProjectile.prefab
@@ -100,7 +100,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   projectileVelocity: 250
-  weapon: {fileID: 8492263470177819836, guid: ba63be8833f2b5d44a70bd02c1a85a99, type: 3}
+  weapon: {fileID: 7485021606706960250, guid: ba63be8833f2b5d44a70bd02c1a85a99, type: 3}
   projectileRb2D: {fileID: 2576749402811358335}
   projectileParticle: {fileID: 1344357935046343875, guid: cf6634ac9202e9a4186449d668feab00, type: 3}
   muzzleFlashParticle: {fileID: 1855075546965395947, guid: ba84c6edddad2734a95382bc309bbaae, type: 3}
