+++ Assets/Prefabs/Projectiles/RapidGunProjectile.prefab
@@ -100,7 +100,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   projectileVelocity: 250
-  weapon: {fileID: 8492263470177819836, guid: ba63be8833f2b5d44a70bd02c1a85a99, type: 3}
+  weapon: {fileID: 2117454392042047476, guid: c65ad5cd03f055e4dbd5e7b644466cc8, type: 3}
   projectileRb2D: {fileID: 2576749402811358335}
   projectileParticle: {fileID: 6609792059005856855, guid: 2f0cce7ba504b4c4bba12b7011703695, type: 3}
   muzzleFlashParticle: {fileID: 5235355868479108926, guid: 935f378eab76c184a837132a598d5d19, type: 3}
