+++ Assets/Scripts/WeaponScripts/ProjectileWeapon.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2d20a0e7fa96d3b41940dbda4c77bb23
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
